+++ src/main/java/me/naming/delieveryservice/config/DatabaseConfig.java
@@ -33,7 +33,7 @@ public class DatabaseConfig {
         SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
         sqlSessionFactoryBean.setDataSource(dataSource);                                    // 위에서 생성된 datasource 빈 객체 주입
         sqlSessionFactoryBean.setVfs(SpringBootVFS.class);                                  // 스프링부트 가상파일 시스템을 설정
-        sqlSessionFactoryBean.setTypeAliasesPackage("me.naming.delieveryservice.vo");       // alias 대상 클래스가 위치한 패키지 경로 설정
+        sqlSessionFactoryBean.setTypeAliasesPackage("me.naming.delieveryservice.dto");       // alias 대상 클래스가 위치한 패키지 경로 설정
         sqlSessionFactoryBean.setMapperLocations(new PathMatchingResourcePatternResolver().getResources("classpath:/mapper/*.xml"));        // Mapper 파일 위치 셋팅
         SqlSessionFactory sqlFactory = sqlSessionFactoryBean.getObject();                   // SqlSessionFactoryBean을 활용해 Mybatis 구성 파일 없이 SqlSessionFactory 빌드
         sqlFactory.getConfiguration().setMapUnderscoreToCamelCase(true);
