+++ src/main/java/me/naming/delieveryservice/dto/RedisPair.java
@@ -1,4 +1,4 @@
-package me.naming.delieveryservice.vo;
+package me.naming.delieveryservice.dto;
 
 
 public class RedisPair {
