+++ src/main/resources/mapper/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="me.naming.delieveryservice.dao.UserDao">
-    <resultMap id="UserVO" type="me.naming.delieveryservice.vo.UserVO">
+    <resultMap id="UserVO" type="me.naming.delieveryservice.dto.UserVO">
         <result property="pid_num" column="pid_num"/>
         <result property="id" column="id"/>
         <result property="password" column="password"/>
